Major refactorings to allow the use of SeverityLevels in FileSetChecks as well as in Checks and to simplify logging of messages in FileSetChecks.
Introduction of a new class called AbstractViolationReporter, from which
both Check and FileSetCheck are derived.

Severity is passed to submodules by means of the Contextualizable
mechanism. This means that the user can f.ex. set severity to INFO
in a TreeWalker and all Checks below that TreeWalker will use severity INFO.